requirements.txt
google-generativeai
pypdf
langchain
matplotlib
networkx
python-dotenv
streamlit
pyvis

